requirements.txt
osrparse==7.0.0
pillow==10.3.0